requirements.txt
Flask==2.3.3
requests==2.31.0
google-cloud-aiplatform
google-cloud-spanner==3.36.0
debugpy
python-dotenv
Flask-Cors
uuid